# analyzeFoodBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyzeFoodBodyGoal } from "./analyzeFoodBodyGoal";

export interface AnalyzeFoodBody {
  /** Base64 encoded image data (without data URL prefix) */
  imageBase64: string;
  /** Image MIME type (e.g. image/jpeg) */
  mimeType: string;
  /** User's dietary goal */
  goal?: AnalyzeFoodBodyGoal;
}