# api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Analyze food image with AI
 */
export const AnalyzeFoodBody = zod.object({
  imageBase64: zod
    .string()
    .describe("Base64 encoded image data (without data URL prefix)"),
  mimeType: zod.string().describe("Image MIME type (e.g. image\/jpeg)"),
  goal: zod
    .enum(["weight_loss", "balanced", "weight_gain"])
    .optional()
    .describe("User's dietary goal"),
});

export const AnalyzeFoodResponse = zod.object({
  foodName: zod.string(),
  calories: zod.number(),
  confidence: zod.number(),
  protein: zod.number().optional(),
  carbs: zod.number().optional(),
  fat: zod.number().optional(),
  suggestion: zod.string().optional().describe("Goal-based dietary suggestion"),
});

/**
 * @summary Get food log entries for a user
 */
export const GetFoodEntriesQueryParams = zod.object({
  date: zod.coerce
    .string()
    .optional()
    .describe("ISO date string (YYYY-MM-DD), defaults to today"),
  userId: zod.coerce.string(),
});

export const GetFoodEntriesResponse = zod.object({
  entries: zod.array(
    zod.object({
      id: zod.number(),
      userId: zod.string(),
      foodName: zod.string(),
      calories: zod.number(),
      confidence: zod.number().optional(),
      protein: zod.number().optional(),
      carbs: zod.number().optional(),
      fat: zod.number().optional(),
      goal: zod.string().optional(),
      imageBase64: zod.string().optional(),
      loggedAt: zod.coerce.date(),
    }),
  ),
  totalCalories: zod.number(),
});

/**
 * @summary Log a food entry
 */
export const CreateFoodEntryBody = zod.object({
  userId: zod.string(),
  foodName: zod.string(),
  calories: zod.number(),
  confidence: zod.number().optional(),
  protein: zod.number().optional(),
  carbs: zod.number().optional(),
  fat: zod.number().optional(),
  goal: zod.string().optional(),
  imageBase64: zod.string().optional(),
});

/**
 * @summary Delete a food log entry
 */
export const DeleteFoodEntryParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteFoodEntryResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get daily calorie summary for a user
 */
export const GetDailySummaryQueryParams = zod.object({
  userId: zod.coerce.string(),
  date: zod.coerce.string().optional(),
});

export const GetDailySummaryResponse = zod.object({
  date: zod.string(),
  totalCalories: zod.number(),
  entryCount: zod.number(),
  goal: zod.string().optional(),
});