# foodEntriesList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FoodEntry } from "./foodEntry";

export interface FoodEntriesList {
  entries: FoodEntry[];
  totalCalories: number;
}