# foodEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FoodEntry {
  id: number;
  userId: string;
  foodName: string;
  calories: number;
  confidence?: number;
  protein?: number;
  carbs?: number;
  fat?: number;
  goal?: string;
  imageBase64?: string;
  loggedAt: Date;
}